backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
langgraph>=0.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
sse-starlette>=2.0.0
